refactoring: add utils folder
