+++ src/utils/addLight.ts
@@ -1,6 +1,6 @@
 import {AmbientLight, Scene} from "three";
 
-export const addLights = (scene: Scene) => {
+export const addLight = (scene: Scene) => {
     const light = new AmbientLight(0x404040, 2);
     scene.add(light);
